# src/main/java/br/com/peoplemanager/dto/address/AddressDto.java
package br.com.peoplemanager.dto.address;

import jakarta.validation.constraints.NotEmpty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
@Builder
public class AddressDto {
    private String id;
    @NotEmpty
    private String street;
    @NotEmpty
    private String zipCode;
    @NotEmpty
    private String number;
    @NotEmpty
    private String city;
    @NotEmpty
    private String state;
    private Boolean principal;
}